requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2 ; python_version >= '3.7'
attrs==23.2.0 ; python_version >= '3.7'
azure-core==1.29.7 ; python_version >= '3.7'
azure-storage-blob==12.19.0
beautifulsoup4==4.12.3 ; python_full_version >= '3.6.0'
brotli==1.1.0
cattrs==23.2.3 ; python_version >= '3.8'
certifi==2023.11.17 ; python_version >= '3.6'
cffi==1.16.0 ; python_version >= '3.8'
chardet==5.2.0 ; python_version >= '3.7'
charset-normalizer==3.3.2 ; python_full_version >= '3.7.0'
cryptography==41.0.7 ; python_version >= '3.7'
cssselect2==0.7.0 ; python_version >= '3.7'
defusedxml==0.7.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django==3.2.23
django-allauth==0.60.1
django-appconf==1.0.6 ; python_version >= '3.7'
django-bootstrap-v5==1.0.11
django-classy-tags==4.1.0 ; python_version >= '3.8'
django-cms==3.11.4
django-compressor==4.4
django-filer==3.1.1
django-formtools==2.5.1 ; python_version >= '3.8'
django-polymorphic==3.1.0
django-sass-processor==1.1
django-sekizai==4.1.0 ; python_version >= '3.8'
django-storages[azure]==1.14.2
django-treebeard==4.7 ; python_version >= '3.8'
djangocms-admin-style==3.2.7 ; python_version >= '3.7'
djangocms-attributes-field==3.0.0
djangocms-file==3.0.1
djangocms-link==3.1.1
djangocms-picture==4.1.1
djangocms-snippet==3.1.1
djangocms-style==3.1.0
djangocms-text-ckeditor==5.1.5
easy-thumbnails[svg]==2.8.5 ; python_version >= '3.6'
exceptiongroup==1.2.0 ; python_version < '3.11'
gunicorn==21.2.0
html5lib==1.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
idna==3.6 ; python_version >= '3.5'
isodate==0.6.1
libsass==0.23.0
lxml==5.1.0 ; python_version >= '3.6'
oauthlib==3.2.2 ; python_version >= '3.6'
packaging==23.2 ; python_version >= '3.7'
pillow==10.2.0 ; python_version >= '3.8'
platformdirs==4.1.0 ; python_version >= '3.8'
psycopg2==2.9.9
pycparser==2.21 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyjwt[crypto]==2.8.0 ; python_version >= '3.7'
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3.post1
rcssmin==1.1.1
reportlab==4.0.9 ; python_version >= '3.7' and python_version < '4.0'
requests==2.31.0
requests-cache==1.1.1
requests-oauthlib==1.3.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rjsmin==1.2.1
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.5 ; python_version >= '3.8'
sqlparse==0.4.4 ; python_version >= '3.5'
svglib==1.5.1 ; python_version >= '3.7'
tinycss2==1.2.1 ; python_version >= '3.7'
typing-extensions==4.9.0 ; python_version < '3.11'
url-normalize==1.4.3 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
urllib3==2.1.0 ; python_version >= '3.8'
webencodings==0.5.1
whitenoise[brotli]==6.6.0
